UPSTREAM: util/apcb: Add apcb_edit tool
On the Picasso architecture, the PSP is responsible for setting up DRAM
before releasing the x86. The APCB (AGESA PSP Configuration Block)
contains multiple SPDs and the GPIO numbers used to select the correct
SPD. Since the source to build the APCBs is not public, it can't be
built as part of the coreboot build. To work around this problem, we use
a template APCB and inject the relevant information.

BUG=b:147042464

Change-Id: Ie3b1dbdfe6cddc9add1d1ff4c6713500f142776e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f836a234e211266d1feffd5f406a692ac0fba164
Original-Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Change-Id: I88a09743f8e8a184c47071ee5e417f5b6bdb7467
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41380
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2123799
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2207106
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>