UPSTREAM: drivers/pc80/rtc: Always load cmos.default if measured boot is enabled
cmos.default used to be loaded only when cmos is needed to be reset,
but conditional loading of CBFS files may change the calculated PCRs
if measurement is hooked on each loading.

In order to resolve this, loadings should be made less conditional,
(if a file might be used, it should be loaded and measured) but the
use of loaded data remains conditional.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1890fee9c8f0c0ad31806e06caf7b789b06cf7cf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 51ce41c0e661fd9cb9207463bcbd920e55b44a62
Original-Change-Id: If6ea0d1cbaa7d96f7dea7e77b7548ca2b30efe9e
Original-Signed-off-by: Bill XIE <persmule@hardenedlinux.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39906
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2131911
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>