UPSTREAM: cros_ec: add chrome EC headers to include path
This adds the path to chrome EC headers to the depthcharge
build. Depthcharge currently includes a manually maintained copy of
the EC headers which are perpetually out of sync with the real
interface definitions. By adding the include path, we can build
depthcharge with the actual EC interface definitions and eliminate the
manual maintenance of copies of EC headers. Once the include path is
in place, we can remove the copies of the EC headers from depthcharge.

BUG=b:152373049

Change-Id: I12946ee1d33f7a88e47972624ee914941e0fb39c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5117c27cc18e6a22e71204d911e404f9e27744e8
Original-Change-Id: I1ce0ad9dc99ea52f177d4fb034fd23efd95a7864
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39947
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2131910
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>