UPSTREAM: libpayload: malloc: Change memcpy() to memmove() in realloc
Our realloc() works (somewhat suboptimally) by free()ing the existing
allocation and then reallocating it wherever it fits. If there was free
space before the old location, this means the new allocation may be
before the old one, and if the free space block is smaller than the old
allocation it may overlap. Thus, we should be moving memmove() instead
of memcpy() to move the block over.

This is not a problem in practice since all our existing memcpy()s are
simple iterate and copy front to back implementations which are safe for
overlaps when the destination is in front of the source. but it's still
the more correct thing to do (in case we ever change our memcpy()s to do
something more advanced or whatever).

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: d61350c403f90ee11c179fece04f68e6c34e1555
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I35f77a94b7a72c01364ee7eecb5c3ff5ecde57f6
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40028
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Change-Id: Ie560a42ab111851640254ee1262344536500e619
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2136780
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>