UPSTREAM: cbfstool: Rename cbfs_walk() to cbfs_legacy_walk()
This function name clashes with cbfs_walk() in the new commonlib CBFS
stack, so rename it to cbfs_legacy_walk(). While we could replace it
with the new commonlib implementation, it still has support for certain
features in the deprecated pre-FMAP CBFSes (such as non-standard header
alignment), which are needed to handle old files but probably not
something we'd want to burden the commonlib implementation with. So
until we decide to deprecate support for those files from cbfstool as
well, it seems easier to just keep the existing implementation here.

BUG=none
BRANCH=none
TEST=none

Change-Id: I34ad0f991e5c243b4831d4d56547a02532985ad1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7066a1e7b3f588c8c4ac0394c8c1f35e227dd552
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I37c7e7aa9a206372817d8d0b8f66d72bafb4f346
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41118
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2577253
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>