UPSTREAM: ec_sync: Run EFS2 in romstage
EFS2 allows EC RO to enable PD for special cases. When doing so, it sets
NO_BOOT flag to avoid booting the OS. AP needs to get NO_BOOT flag from
Cr50 and enforce that.

This patch makes verstage get a boot mode and a mirrored hash stored
in kernel secdata from Cr50.

This patch also makes romstage write an expected EC hash (a.k.a. Hexp) to
Cr50 (if there is an update).

BUG=b:147298634, chromium:1045217, b:148259137
BRANCH=none
TEST=Verify software sync succeeds on Puff.

Signed-off-by: Dossym Nurmukhanov <dossym@chromium.org>
Original-Commit-Id: dff56a056c7dadf0d970cfe29f0bc9c1fec69e82
Original-Signed-off-by: dnojiri <dnojiri@chromium.org>
Original-Change-Id: I1f387b6e920205b9cc4c8536561f2a279c36413d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40389
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Change-Id: I356dddedc39c1ca1fe256544f196e56acb58b3ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2155377