UPSTREAM: mb/google/volteer: Work around TPM issue by enabling GPIO PM in S0ix
Setting the default values for GPIO community power management, causes
issues in detecting TPM interrupts. So to avoid that GPIO PM has to be
disabled in devicetree. But for S0ix it is needed. This patch implements
a workaround in ASL code to enable GPIO PM on S0ix entry and disable it
on S0ix exit.

This patch adds the following three platform specific methods.

1. MS0X to enable power management features for GPIO communities on
entry and on exit, it disables them.

2. MPTS to enable power management features for GPIO communities when
preparing to sleep.

3. MWAK to disable power management features for GPIO communities on
waking up.

BUG=b:148892882
BRANCH=none
TEST="Boot with this change on volteer proto1 and check for GPIO
community config with debugger"

Change-Id: If8e0be8fb33649d47ba9dc05982cff05a3674a41
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: eb148d88d1af5b4fb409794e974f89cf2705741a
Original-Signed-off-by: Venkata Krishna Nimmagadda <venkata.krishna.nimmagadda@intel.com>
Original-Change-Id: If522c82c0069a4bf5738beb73a2b4f11ed6f51d3
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40261
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Original-Reviewed-by: Venkata Krishna Nimmagadda <Venkata.krishna.nimmagadda@intel.com>
Original-Reviewed-by: Wonkyu Kim <wonkyu.kim@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2173034
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>