UPSTREAM: trogdor: Add third RAM_CODE pin
We decided to add a third RAM_CODE pin to the Trogdor family for devices
after rev1. This patch adds support to read it. Since the newly used pin
was previously unconnected (not pulled down) on rev1, this will change
the RAM_CODE result for previous versions (and actually make it
undetermined until we enable tri-state). But since we're not actually
using RAM_CODE for anything yet, and since those are development
revisions that will eventually be discontinued, this should be fine.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib0a198a950755c56b34473580e70f7dd4dbd3ab3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ef43711aadcaf61ed1234035afc2e0282bb94e97
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I9b52982f17646a305b1a3e2c7d37606a7c38d0c4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40352
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Philip Chen <philipchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2151593
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>