UPSTREAM: external/LinuxBoot: Kconfig defaults to systemboot
In addition to change CB:40316 this commit proposes a change of default
behavior inside of the LinuxBoot toolchain.

Currently the defaults build a LinuxBoot payload which boots into the
u-root shell and waits for input. In fact it does not deliver any
bootloader with it, but the build image is on the other hand rather
small.

This commit changes the defaults in a way that the LinuxBoot
bootloader will actually be able to boot a local or remote system image.
In consequence the build payload size is rather big, but accepted for a
working out-of-the-box startup behavior.

See discussion on that topic in the above mentioned change.

(cherry picked from commit ed8aef9ccdfd909af25dcf412b1b4dce4c104fbc)

Original-Signed-off-by: Patrik Tesarik <mail@patrik-tesarik.de>
Original-Change-Id: Ieaba7e523aef10c467a8bea29ae323e22324b225
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40527
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: ed8aef9ccdfd909af25dcf412b1b4dce4c104fbc
Change-Id: I0e66b0b2fa167140e37288391c9750c1a0491b1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3707198
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>