UPSTREAM: mb/lenovo/x200: Add support for ThinkLight
With this patch, the ThinkLight on the ThinkPad X200 can be controlled
through the OS.  This was initially done for the X201 in f63fbdb6:
mb/lenovo/x201: Add support for ThinkLight.

After applying this patch, the light can be controlled like this:

    echo on >/proc/acpi/ibm/light
    echo off >/proc/acpi/ibm/light

Or through sysfs at /sys/class/leds/tpacpi::thinklight

I have tested it on an X200 with Kernel 5.4 and it seems to work fine.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id46d75df48c57c5b3c8f67ddd89176b755d570b1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b45912f453b52bfc01e66dc070fb33fad236d816
Original-Change-Id: I14752ab33484122248959517e73f96b6783b1f65
Original-Signed-off-by: Stefan Ott <stefan@ott.net>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40620
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2170947
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>