UPSTREAM: mb/lenovo/x220: Add support for ThinkLight
With this patch, the ThinkLight on the ThinkPad X220 can be controlled
through the OS.  This was initially done for the X201 in f63fbdb6:
mb/lenovo/x201: Add support for ThinkLight.

After applying this patch, the light can be controlled like this:

    echo on >/proc/acpi/ibm/light
    echo off >/proc/acpi/ibm/light

Or through sysfs at /sys/class/leds/tpacpi::thinklight

Unfortunately I do not own an X220 to test this.

BUG=none
BRANCH=none
TEST=none

Change-Id: I95e0a4a29fda43c25e5c730b5e053da627c04a4f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: be698de76e102fd7801444a04343c4bfbad1ff2a
Original-Change-Id: Icead793694475e2f63353690203790ab7ce7c597
Original-Signed-off-by: Stefan Ott <stefan@ott.net>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40668
Original-Reviewed-by: Alexander Couzens <lynxis@fe80.eu>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2216010
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>