UPSTREAM: soc/amd/picasso: Drop addition of PSPTRUSTLETS_FILE
PSPTRUSTLETS_FILE was including a binary for fTPM which according to
BIOS architecture design guide is the firmware enabled TPM. Chrome OS
does not really use firmware enabled TPM. Also, this is an option
which is mainboard dependent.

This change drops the addition of PSPTRUSTLETS_FILE to PSP
directory. If this is something that is required by any mainboard,
there should be a separate Kconfig to include the required files.

BUG=b:154880818
TEST=Verified that trembyle still boots

Change-Id: I18ee8e8c524681ab7de2ce70f7326e2547ee3dd3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4e8b63970382d18498216ee84f3c9ee3bf3ec67d
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: Iaa2126c879986d00c921c85fb5cb5257c7065006
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40685
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2173052
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>