UPSTREAM: nb/intel/haswell/pei_data.h: Add ULT system type
Looks like 5 is a valid system type, as Google Beltino and Slippy are
using it. According to comments on these mainboards' code, this value
corresponds to ULT systems. So, add it to the comment on the pei_data
struct, which was likely copied from Sandy Bridge and was not updated.

BUG=none
BRANCH=none
TEST=none

Change-Id: I819b5dc20da613d68f74db3c7f0eda67ab0cbd16
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6ba3a0758f7162834378e5845ea3a65fe327012c
Original-Change-Id: I3654bb6022839dba3e1499cf43e8beaa97d1def1
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40692
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2175815
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>