UPSTREAM: acpi_device: Make integer array input variable const
An array of 64bit integers is passed to acpi_dp_add_integer_array() but
it is not const so can't take a const array without a compiler error.

The function does not modify the array so it can be made const without
breaking anything and allowing a const array to be passed in the future.

BUG=b:146482091

Change-Id: I2395a97de913632057db02d4156b2e4323268212
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ed6eb2713a738c45ab876941d6b5805fd00bf13c
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I98ecdaef5ddfa2026390e2812f5ea841ee51f073
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40882
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2175881
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>