UPSTREAM: device: Add definitions for SoundWire specification
This header implements structures to describe the properties defined in
the SoundWire Discovery and Configuration Specification Version 1.0.

By itself this just provides the property definitions, it is then used
by the code that generates ACPI device properties and by the controller
and codec drivers.

A new header for MIPI vendor/device IDs is also added, with the MIPI
Alliance board members added by default.  This will be used in the same
way as pci_ids.h to track devices added to coreboot.

BUG=b:146482091

Change-Id: Ia72d60b332501634efa98c2fb0968f64d1b4164c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4480dc1ca1fd6982a3b4cbc6cf1531992f517fb2
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: Ie9901d26d1efe68edad7c049c98a976c4e4f06f4
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40883
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2209469
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>