UPSTREAM: soc/intel/tigerlake: Add definition for PMC EPOC
The PMC EPOC register indicates which external crystal oscillator is
connected to the PCH.  This frequency is important for determining the
IP clock of internal PCH devices.

Add definitions that allow this register to be read and extract the
crystal frequency, and a helper function to extract and return this
as the defined enum.

BUG=b:146482091

Change-Id: Ia0a2d94b02c13ed0522055df1c4392ce8211f11a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 526880754faa1246e97e0bd1d836ecc1738e8c90
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I959fe507f3dbf93b6176b333a9e725ed09f56328
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40887
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2216000
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>